simplify display flag code
