+++ liwords-ui/src/lobby/sought_games.tsx
@@ -12,8 +12,7 @@ import {
 import { SoughtGame } from '../store/reducers/lobby_reducer';
 import { useMountedState } from '../utils/mounted';
 import { PlayerAvatar } from '../shared/player_avatar';
-import { DisplayFlag } from '../shared/display_flag';
-import { useBriefProfile } from '../utils/brief_profiles';
+import { DisplayUserFlag } from '../shared/display_flag';
 import { RatingBadge } from './rating_badge';
 
 export const timeFormat = (
@@ -48,7 +47,6 @@ type PlayerProps = {
 };
 
 export const PlayerDisplay = (props: PlayerProps) => {
-  const profile = useBriefProfile(props.userID);
   return (
     <div className="player-display">
       <PlayerAvatar
@@ -58,7 +56,7 @@ export const PlayerDisplay = (props: PlayerProps) => {
         }}
       />
       <span className="player-name">{props.username}</span>
-      <DisplayFlag countryCode={profile?.getCountryCode()} />
+      <DisplayUserFlag uuid={props.userID} />
     </div>
   );
 };
