+++ liwords-ui/src/shared/usernameWithContext.tsx
@@ -2,10 +2,9 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { Dropdown } from 'antd';
 import { TheBlocker } from './blocker';
-import { useBriefProfile } from '../utils/brief_profiles';
 import { useLoginStateStoreContext } from '../store/store';
 import { canMod } from '../mod/perms';
-import { DisplayFlag } from './display_flag';
+import { DisplayUserFlag } from './display_flag';
 import { SettingOutlined } from '@ant-design/icons';
 import { TheFollower } from './follower';
 
@@ -32,7 +31,6 @@ type UsernameWithContextProps = {
 export const UsernameWithContext = (props: UsernameWithContextProps) => {
   const { loginState } = useLoginStateStoreContext();
   const { loggedIn, userID, perms } = loginState;
-  const briefProfile = useBriefProfile(props.userID);
   const userMenu = (
     <ul>
       {loggedIn &&
@@ -113,9 +111,7 @@ export const UsernameWithContext = (props: UsernameWithContextProps) => {
         ) : (
           <>
             {props.fullName || props.username}
-            {briefProfile && props.includeFlag && (
-              <DisplayFlag countryCode={briefProfile!.getCountryCode()} />
-            )}
+            {props.includeFlag && <DisplayUserFlag uuid={props.userID} />}
           </>
         )}
       </span>
